+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Simpler mq.Client interface with one Send method
 - tt.Client uses stack of middlewares
 - Add mq.Middleware interface
 - Add tt.LogLevel with info and debug